Metadata reorganization
- prjman.metadata.base -> prjman.metadata
- prjman.metadata.* -> prjman.meta.*
