+++ src/prjman/cli.py
@@ -3,7 +3,7 @@
 
 import click
 import prjman.workspace.project as wspc
-from prjman.metadata.base import ProjectMetadata
+from prjman.metadata import ProjectMetadata
 from prjman.config import PrjmanConfig, load_config
 
 @click.group()
