+++ src/prjman/meta/files/osf.py
@@ -4,7 +4,7 @@ project.
 
 from prjman.module import SINGLETON
 from prjman.struct.codec import DictObject
-from prjman.metadata.file import ProjectFile, ProjectFileCodec, build_file
+from prjman.meta.file import ProjectFile, ProjectFileCodec, build_file
 from prjman.utils import download_and_write
 
 class OSFProjectFile(ProjectFile):
